+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-## 3.0.8
+## 3.0.9
 
 ### Fixed
 - ResetPasswordController, RegisterController, ProfileController change response type
 - ElementsTrait fix for Blog Feed
 - uk-list-striped fix
 - Prettierrc
+- Navbar title translate
+- Latin characters issue fixed
 
 ### Added
 - New Icon: key.svg